Rename @linky/core to @linky/identity
