+++ apps/push/Dockerfile
@@ -8,7 +8,7 @@ COPY apps/native-shell/package.json apps/native-shell/package.json
 COPY apps/push/package.json apps/push/package.json
 COPY apps/site/package.json apps/site/package.json
 COPY apps/web-app/package.json apps/web-app/package.json
-COPY packages/core/package.json packages/core/package.json
+COPY packages/identity/package.json packages/identity/package.json
 COPY packages/config/package.json packages/config/package.json
 COPY packages/linkstr/package.json packages/linkstr/package.json
 COPY packages/linkstr-react/package.json packages/linkstr-react/package.json
