+++ apps/web-app/package.json
@@ -24,7 +24,7 @@
     "@evolu/common": "^7.4.1",
     "@evolu/react": "^10.4.0",
     "@evolu/react-web": "^2.4.0",
-    "@linky/core": "workspace:*",
+    "@linky/identity": "workspace:*",
     "@linky/linkstr": "workspace:*",
     "@linky/linkstr-react": "workspace:*",
     "@noble/curves": "^1.4.1",
