+++ packages/identity/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@linky/core",
+  "name": "@linky/identity",
   "private": true,
   "version": "0.1.0",
   "license": "0BSD",
