+++ packages/identity/src/identity/MasterSecretProvider.ts
